Improve the accessibility of DownloadButton in playlists
Previously, the download button was nested inside another button for the
playlist row which caused an accessibility error. This commit moves the
download button outside of the playlist item button but positions it in
the same position as it was previously on desktop/mobile. The user can
tab to the button and it will be announced to screen readers.

Additionally, this change adds 8px padding to the download button in
playlists on mobile as I found it was difficult to tap on the button.