Reinstate the video poster image when playing audio sources
If the video player loads a video we want its poster image to be the
first frame of the video. However, if the video player loads an audio
source then we don’t want to show a black screen and should instead show
the advert/content image as we did previously.