Avoid loading styles in some cases, e.g. when playerStyle: “none”
Previously, the UserInterface contained a conditional for deciding
whether its root element should render or not, but we had already loaded
the styles by this point because the import line is at the top of the
UserInterface component. Instead, we should move this conditional out
to the Player component so that we avoid loading styles in some cases.