Don’t show the video poster if findLoadedMedia failed
Previously, we were failing to identify loadedMedia and were therefore
showing the video poster over the top of the video player, even though
the video had actually loaded. I think in this case, we shouldn’t show
the poster because we don’t know whether or not the loaded source is
video and I’d rather try to show it if it is available instead of
covering it up by default with the poster image.