Fix a problem when multiple script tags are added to the page
If the player script was loaded multiple times, the browser would only
fetch the script once and execute it once. However, the result is then
cached and the window.BeyondWords global was being redeclared each time
because it was the named export in the vite.config.ts build config.

This meant that static data stored on BeyondWords.Player was being lost
which included the #instances variable. As a result, when the
styleLoaded function was called from the style.js file, it wasn’t
clearing the ‘display: none’ from all of the instances and they were
remaining hidden.

Therefore, set a different name for the export in vite.config.js so that
we don’t clash with window.BeyondWords which is already defensively
handled at the bottom of the index.ts file. This will ensure the same
class declaration is used for all player instances.