Fix the Hls library not loading when multiple script tags are added to the page
I’m not sure why, but the previous approach was only setting Hls for one
of the player instances except on rare occasions where you would refresh
and it would set the Hls library for all instances (work correctly).

I think it might be something to do with the combination of the ||=
operator and the ‘await import’ not working well together. This approach
now works for 100% of page refreshes with the downside that it clobbers
the window.Hls global if it has already been set.