Make (xpath, md5) a composite key for matching purposes
Previously we were matching segments on marker, xpath, md5 independently
but I think it will be more correct if we have to match on both xpath
and md5 together if no marker is supplied. This way, we won’t mistakenly
highlight segments on the page if the xpath matches but the content is
not what we are expecting it to be.