Fix the md5Cache not working correctly
If two different paragraphs were used as keys to the object they were
returning the same value. WeakMap does not have this issue and has the
added benefit that it does not keep DOM memory alive if there are no
other references to the DOM node (i.e. if it is removed).