Add the segment widget next to the first matching element
Previously, we were always adding it next to the segment.segmentElement
but we want it to appear next to the first segment on the page that
has the same marker (if one exists). Otherwise, this structure would
cause the widget to appear next to the first element that was hovered
with the mouse rather than the first list item which is preferred.

<ul data-beyondwords-marker=“foo”>
  <li data-beyondwords-marker=“foo”>List item 1</li>
  <li data-beyondwords-marker=“foo”>List item 2</li>
</ul>