Ignore PressedSegment events for precedence > 0
If another player takes precedence over the current one for handling
the PressedSegment event then ignore the event. Otherwise, multiple
players will try to play the segment and pause each other causing a kind
of deadlock.

Although we aren’t currently using precedence in HoveredSegmentUpdated
and CurrentSegmentUpdated, we do calculate this in chooseSegmentPerPlayer.ts
and it might be useful later for other kinds of tie-breaks so ensure
this field is included in the MediaElement’s CurrentSegmentUpdated event
for consistency. In this case, it will always be 0 because it is emitted
from the singular player that is playing.