Defer loading the first .ts fragment until played
Split loadMedia into two phases so that we load the .m3u8 metadata on
page load but don’t try to load any audio/video sources until the user
presses play. This reduces the number of requests to CloudFront.

Maybe in the future we could look to defer the .m3u8 fetch until play is
pressed as well, but that is a bigger refactor.

This commit introduces a new MetadataLoaded event which serves as a
replacement for anyone currently relying on the MediaLoaded event to
detect when the player has loaded.