Fix intros/outros not playing
We recently changed the workflow when playing audio. It now loads the
media when pressing play so that we don’t unnecessarily request media
files until they’re used.

However, this means the load happens when you press play and if there is
an intro outro then we immediately change the media source and start to
load that instead, which then results in an AbortError because the first
play call hasn’t finished yet.

If this happens, I think we should just retry since the user tried to
play some audio and it didn’t work because a new load occurred.