Fix the player getting stuck if the same media is played twice in a row
On the player demo page, we make a playlist of the same item 5 times:
https://beyondwords-io.github.io/player-demo/

The MediaElement component assumes that each time loadMetadata is called,
the onMetadata callback will be called (eventually). However, if we are
not using HLS and are relying on the <source> element to load the media,
we don’t receive another metadatadaLoaded event because the source
hasn’t changed.

To fix this, we should satisfy the assumption that MediaElement is
making about loadMetadata and manually call onMetadata({}) ourselves.
The handler function in MediaElement ensures that we don’t send a
duplicate MetadataLoaded player event which seems correct since the
video element isn’t emitting one. We only want to do this if the video
already has a sourceUrl so that we don’t prematurely call onMetadata for
cases when the video element will emit the event for us.