Fix outros not playing
We need to make sure we nullify the hls variable in MediaElement if an
mp3 file is loaded (e.g. for the outro). Otherwise, this line is being
evaluated incorrectly at the top of handleLoadedMetadata:

const hlsBeingUsed = hls !== "not-used";

This was then preventing the MetadataLoaded event for outros because it
assumes the handleLoadedMedia function would have already been called by
the HLS manifestLoaded event handler.