Encode the sourceUrl when requesting /player/by_source_url
Previously, we were sending an unencoded request to:

/player/by_source_url/https://example.com

The Rails route that matches this URL was interpreting the double slash
as a single slash which is a common convention in routing. This meant
that the X-Referer header did not match the URL from the path and the
client-side integration workflow wasn’t running.

Instead, make sure the URL is encoded when making the request

/player/by_source_url/https%3A%2F%2Fexample.com

This ensures Rails parses the :url route parameter correctly.