Fix the build failing the first time after git cloning
The daisybitStrings.ts file was previously being via a Vite plugin but
this was unreliable because it would not exist the first time you run
the ./bin/build script after git cloning the project. This was causing
an import error. Instead, add a ./bin/write_daisybits script and run it
explicitly before the ./bin/server and ./bin/build scripts.