Fix the player crashing when clicking on segments across playlist items
I’m not really sure why this fixes the problem. I tried to set up the
Svelte project so that I could fix the bug properly rather than having
to use tick but I wasn’t able to get it set up for development despite
trying for 1-2 hours. It also looks like this implementation has
significantly changed in Svelte 5 so I am hopeful that will mean the
workaround is no longer needed when that releases.

https://linear.app/beyondwords/issue/S-3849/fix-a-player-crash-when-clicking-on-segments-across-playlist-items

https://github.com/beyondwords-io/player/commit/cd61b7a7b9f6da8de4851053cebee0e54fa9b3d5

https://github.com/sveltejs/svelte/issues/4471