Allow play() to be called without metadataLoaded on iOS Safari
https://linear.app/beyondwords/issue/S-4400/player-not-loaded-in-ios-wkwebview-when

The MetadataLoaded event isn’t emitted until a user interaction occurs
on iOS Safari if the player is initialised inside of a setTimeout. In
this case, we should bypass the check that the metadata has loaded and
allow playback to begin. If the metadata fails to load for some reason
then we should receive a JavaScript error, but this would be unexpected
since the .m3u8 URL in the API response should always work.