Fix VAST adverts immediately re-playing after pressing pause
https://linear.app/beyondwords/issue/S-4405/fix-a-vast-advert-not-being-pausable-for-nine-australia

For some reason, the value of video.paused is always false (i.e. playing
is true) while the Google IMA SDK is taking over playback, even when the
adManager is also paused. This seems like an IMA SDK bug.

This incorrect value was triggering some code added in commit edd4ec58
that tries to keep the UI in sync with the video element, for example,
if video.pause() and video.play() are called directly on the underlying
element, bypassing the playbackState property of the player.

To fix this, we can always assume the video is paused when a VAST advert
is playing. This is technically true because the Google IMA SDK adds an
iframe which is the element that is actually playing, not the video.