Match segments on md5 only if the xpath match fails
This can happen if the content that was sent to the BeyondWords API is
different from the the DOM tree on the source page. For example, if the
publisher removed HTML boilerplate elements (body, article, etc.) from
the HTML when importing it into the system.

This can also happen if there’s some JavaScript that modifies the source
page’s DOM structure, e.g. adding the player itself can cause indexes
to change in xpath, e.g. /html/body/p[1] might become /html/body/p[2].