Set matchedElement on segments (including siblings of the current node)
When playback advances to the next segment, we try to identify the
corresponding DOM node so we can highlight it. We try to identify the
node by data-beyondwords-marker or xpath, but this might not always work,
for example, if the DOM tree of content in the API doesn't match the
source page.

In chooseSegmentPerPlayer, we fall back to matching on MD5 in response
to the user hovering or clicking on a segment, but this does not handle
the case when playback advances since there will be no 'target'
available in that context. This would mean we'd need to scan the entire
DOM to identify a node by MD5 when playback advances. Even if we did
this once ahead of time, it would be expensive.

Therefore, we work around this issue by only scanning sibling elements
of nodes that were hovered or clicked, rather than the entire DOM. This
enables highlighting of segments to work as playback advances, provided
the user has moved their mouse over one of the segments on the page
first. This isn't a perfect solution but seems like a reasonable
compromise given the alternative would be expensive.