Don’t keep requesting the manifest file if the request 404s
If we try to fetch the .m3u8 manifest file and receive a 404, previously
we would set hls = null which then triggered loadMetadata again because
of Svelte’s reactivity. We’d then request the file again and fatal error
again over and over again with very little delay between requests which
would result in a lot of unnecessary requests being made to CloudFront.