Fix the player freeze framing on HLS videos in Android version <= 13
HLS support in Android version 13 and below is not very good. We already
have to use the mp4 media instead of the m3u8 media on Android if an
initial time is set (time fragments) and it looks like this is another
case where we should load the mp4 instead of m3u8 streaming format.

https://linear.app/beyondwords/issue/S-5272/video-player-freezing-on-android-mobile-words-stuck-audio-continues

https://lstnltd.slack.com/archives/C061NQDK1SL/p1715774612024949