Don’t skip the intro if a user clicks on a segment
Previously, we treated the intro kind of like a playlist track the comes
immediately before the content audio, but we now want to play the intro
anyway if a user clicks on a segment. They can still skip past it unlike
adverts and the advert will still play after the intro if there is one.